requirements.txt
pyicloud
pyyaml